refactor(rest-api-protocol): restructure v1 resources into context-based modules
- Rename v1/resources/ to v1/context/ with per-context subdirectories
- Each context has a resources/ subdirectory for DTO files
- Rename health_check/resources/health_check.rs to report.rs
- Update all Rust imports across the workspace
- Update EPIC and sub-issue specs to document the convention
