+++ packages/axum-rest-api-server/src/v1/context/health_check/handlers.rs
@@ -2,7 +2,7 @@
 //! API context.
 
 use axum::Json;
-use torrust_tracker_rest_api_protocol::v1::resources::health_check::{Report, Status};
+use torrust_tracker_rest_api_protocol::v1::context::health_check::resources::report::{Report, Status};
 
 /// Endpoint for container health check.
 pub async fn health_check_handler() -> Json<Report> {
