+++ packages/axum-rest-api-server/src/v1/context/torrent/responses.rs
@@ -2,17 +2,17 @@
 //! API context.
 use axum::response::{IntoResponse, Json, Response};
 use serde_json::json;
-use torrust_tracker_rest_api_protocol::v1::resources::torrent::{ListItem, Torrent};
+use torrust_tracker_rest_api_protocol::v1::context::torrent::resources::torrent::{ListItem, Torrent};
 
 /// `200` response that contains an array of
-/// [`ListItem`](torrust_tracker_rest_api_protocol::v1::resources::torrent::ListItem)
+/// [`ListItem`](torrust_tracker_rest_api_protocol::v1::context::torrent::resources::torrent::ListItem)
 /// resources as json.
 pub fn torrent_list_response(items: Vec<ListItem>) -> Json<Vec<ListItem>> {
     Json(items)
 }
 
 /// `200` response that contains a
-/// [`Torrent`](torrust_tracker_rest_api_protocol::v1::resources::torrent::Torrent)
+/// [`Torrent`](torrust_tracker_rest_api_protocol::v1::context::torrent::resources::torrent::Torrent)
 /// resources as json.
 pub fn torrent_info_response(torrent: Torrent) -> Json<Torrent> {
     Json(torrent)
