+++ packages/axum-rest-api-server/tests/server/v1/contract/context/health_check.rs
@@ -1,6 +1,6 @@
 use torrust_tracker_axum_rest_api_server::testing::environment::Started;
 use torrust_tracker_rest_api_client::v1::client::get;
-use torrust_tracker_rest_api_protocol::v1::resources::health_check::{Report, Status};
+use torrust_tracker_rest_api_protocol::v1::context::health_check::resources::report::{Report, Status};
 use torrust_tracker_test_helpers::{configuration, logging};
 use url::Url;
