+++ packages/rest-api-application/src/ports/torrent.rs
@@ -2,7 +2,7 @@
 use async_trait::async_trait;
 use torrust_info_hash::InfoHash;
 use torrust_tracker_primitives::pagination::Pagination;
-use torrust_tracker_rest_api_protocol::v1::resources::torrent::{ListItem, Torrent};
+use torrust_tracker_rest_api_protocol::v1::context::torrent::resources::torrent::{ListItem, Torrent};
 
 /// Port for querying torrent data from the tracker runtime.
 ///
