+++ packages/rest-api-protocol/src/v1/context/health_check/mod.rs
@@ -1,0 +1,5 @@
+//! Health check context — `/api/health_check` endpoint.
+//!
+//! Refer to the [`axum-rest-api-server`] counterpart at
+//! `v1::context::health_check` for the HTTP routing and handler layer.
+pub mod resources;
