+++ packages/rest-api-protocol/src/v1/context/health_check/resources/mod.rs
@@ -1,0 +1,2 @@
+//! Resources for the [`health_check`](super) context.
+pub mod report;
