+++ packages/rest-api-protocol/src/v1/context/mod.rs
@@ -1,0 +1,6 @@
+//! API resources (DTOs) for the v1 REST API contract, organized by context.
+//!
+//! Each submodule corresponds to an API context. Resources for each context
+//! live under its `resources/` subdirectory.
+pub mod health_check;
+pub mod torrent;
