+++ packages/rest-api-protocol/src/v1/context/torrent/mod.rs
@@ -1,0 +1,2 @@
+//! Torrent context — protocol DTOs for the torrent API group.
+pub mod resources;
