+++ packages/rest-api-protocol/src/v1/context/torrent/resources/mod.rs
@@ -1,0 +1,3 @@
+//! Resources for the [`torrent`](super) context.
+pub mod peer;
+pub mod torrent;
