--- packages/rest-api-protocol/src/v1/resources/mod.rs
@@ -1,6 +1,0 @@
-//! API resources (DTOs) for the v1 REST API contract.
-//!
-//! Each module corresponds to a resource group in the API.
-pub mod health_check;
-pub mod peer;
-pub mod torrent;
