+++ packages/rest-api-runtime-adapter/src/conversion.rs
@@ -4,8 +4,8 @@
 //! `BasicInfo`, `peer::Peer`) with transport-agnostic protocol DTOs.
 use torrust_tracker_core::torrent::services::{BasicInfo, Info};
 use torrust_tracker_primitives::{PeerId, peer as domain_peer};
-use torrust_tracker_rest_api_protocol::v1::resources::peer as protocol_peer;
-use torrust_tracker_rest_api_protocol::v1::resources::torrent::{ListItem, Torrent};
+use torrust_tracker_rest_api_protocol::v1::context::torrent::resources::peer as protocol_peer;
+use torrust_tracker_rest_api_protocol::v1::context::torrent::resources::torrent::{ListItem, Torrent};
 
 /// Convert a domain [`domain_peer::Peer`] into a protocol [`protocol_peer::Peer`].
 #[must_use]
