+++ src/console/ci/qbittorrent_e2e/scenario_steps/tracker/verify_tracker_swarm.rs
@@ -1,5 +1,5 @@
 use anyhow::Context;
-use torrust_tracker_rest_api_protocol::v1::resources::torrent::Torrent;
+use torrust_tracker_rest_api_protocol::v1::context::torrent::resources::torrent::Torrent;
 
 use super::super::super::tracker::TrackerApiClient;
 use super::super::super::types::InfoHash;
